requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.0
pymongo[srv]==4.6.1
Pillow==10.2.0
certifi==2024.2.2

